pyproject.toml
[project]
name = "ecommercereinabackend"
version = "0.1.0"
description = "backend-django"
authors = [
    {name = "Sebastian Cerquera",email = "sebartin9@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg (>=3.2.13,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "python-semantic-release (>=10.5.2,<11.0.0)"
]

[tool.semantic_release]
version_toml = [
    "pyproject.toml:project.version"
]
branch = "main"
upload_to_pypi = false
upload_to_release = true
build_command = "echo 'No build needed'"
commit_message = "chore(release): {version} [skip ci]"
